"refactor(package): Refactored the file structure of the repository and add CI"
