+++ .github/workflows/gitflow-R-action.yml
@@ -1,0 +1,16 @@
+name: Gitflow Action for R Package Development
+run-name: ${{ github.actor }} now ${{ github.event_name }} to ${{ github.ref_name }}
+
+on:
+  push:
+    branches: ['*', '!dev', '!main']
+    
+  pull_request:
+    types: ['opened', 'closed']
+    branches: ['main', 'release*', 'master']
+
+jobs:
+  Activating_Parser:
+    uses: fnlcr-bids-sdsi/gitflow-R/.github/workflows/parser.yml@single-cell-analysis
+    with:
+      image_to_use: "ghcr.io/ruiheesi/pyscworkflow:latest"
