+++ tests/test_hierarchical_heatmap.py
@@ -2,7 +2,7 @@ import sys
 import os
 sys.path.append(os.path.dirname(os.path.realpath(__file__)) + "/../src")
 
-from scworkflow.hierarchical_heatmap_module import hierarchical_heatmap
+from scsaworkflow.visualization import hierarchical_heatmap
 import anndata
 import pandas as pd
 import numpy as np
